requirements.txt
aggdraw==1.3.18
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
black==23.12.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
docopt==0.6.2
fastapi==0.104.1
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.46.0
frozenlist==1.4.1
fsspec==2023.12.1
gradio==3.48.0
gradio_client==0.6.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
imageio==2.33.1
importlib-resources==6.1.1
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.8.2
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.2
opencv-python==4.8.1.78
orjson==3.9.10
packaging==23.2
pandas==2.1.4
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.1.0
psd-tools==1.9.28
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rpds-py==0.13.2
scikit-image==0.22.0
scipy==1.11.4
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tifffile==2023.12.9
toolz==0.12.0
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
websockets==11.0.3
yarl==1.9.4
